+++ src/azure-cli/azure/cli/command_modules/resource/tests/latest/test_resource.py
@@ -26,9 +26,10 @@ class ResourceGroupScenarioTest(ScenarioTest):
         self.cmd('group exists -n {rg}',
                  checks=self.check('@', False))
 
-        self.cmd('group create -n {rg} -l westus --tag a=b c', checks=[
+        self.cmd('group create -n {rg} -l westus --tag a=b c --managed-by test_admin', checks=[
             self.check('name', '{rg}'),
-            self.check('tags', {'a': 'b', 'c': ''})
+            self.check('tags', {'a': 'b', 'c': ''}),
+            self.check('managedBy', 'test_admin')
         ])
         self.cmd('group exists -n {rg}',
                  checks=self.check('@', True))
@@ -65,8 +66,10 @@ class ResourceGroupNoWaitScenarioTest(ScenarioTest):
                  checks=self.is_empty())
         self.cmd('group exists -n {rg}',
                  checks=self.check('@', False))
-        self.cmd('group create -n {rg} -l westus',
-                 checks=self.check('name', '{rg}'))
+        self.cmd('group create -n {rg} -l westus --managed-by test_admin', checks=[
+            self.check('name', '{rg}'),
+            self.check('managedBy', 'test_admin')
+        ])
         self.cmd('group exists -n {rg}',
                  checks=self.check('@', True))
         self.cmd('group wait --exists -n {rg}',